# src/main/resources/api.yml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Medical Image Analysis API
  description: REST API Documentation
  termsOfService: https://medical-image-analysis.com/terms-of-use
  contact:
    name: Medical Image Analysis API Support
    email: support@medical-image-analysis.com
servers:
  - url: https://api.medical-image-analysis.com/v1
    description: Production server
  - url: https://api-dev.medical-image-analysis.com/v1
    description: Internal staging server for testing
paths:
  /api/v1/users/{userId}:
    get:
      tags:
        - Users
      summary: Return User
      operationId: getUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Data of specified user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              examples:
                user:
                  $ref: '#/components/examples/user'
        404:
          $ref: '#/components/responses/UserNotFound'
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Return Users
      operationId: getUsers
      responses:
        200:
          description: Basic Data users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BasicUser'
              examples:
                basicUsers:
                  $ref: '#/components/examples/basicUsers'
        404:
          $ref: '#/components/responses/UserNotFound'
  /api/v1/users/{userId}/medical-images:
    get:
      tags:
        - Users
      summary: Return User medical images
      operationId: getUserMedicalImages
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: List of specified user medical images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicalImageResponse'
              examples:
                medicalImageResponseExample:
                  $ref: '#/components/examples/medicalImagesResponseExample'
    post:
      tags:
        - Users
      summary: Create medical images
      operationId: createMedicalImages
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        description: Medical images
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MedicalImage'
      responses:
        201:
          description: List of specified user medical images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalImageResponse'
              examples:
                medicalImageResponseExample:
                  $ref: '#/components/examples/medicalImageResponseExample'
  /api/v1/users/{userId}/medical-images/{imageId}:
    get:
      tags:
        - Users
      summary: Return user medical image
      operationId: getMedicalImage
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Get medical image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalImageResponse'
              examples:
                medicalImageResponseExample:
                  $ref: '#/components/examples/medicalImageResponseExample'
  /api/v1/medical-images:
    get:
      tags:
        - Medical Images
      summary: Return medical images
      operationId: getMedicalImages
      responses:
        200:
          description: List medical images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicalImageResponse'
              examples:
                medicalImagesResponseExample:
                  $ref: '#/components/examples/medicalImagesResponseExample'
components:
  examples:
    basicUsers:
      summary: List of users
      value:
        [
        {
          id: 1,
          username: jeremiasz.jaworski,
          firstName: Jeremiasz,
          lastName: Jaworski,
          email: jeremiasz.jaworski@outlook.com,
          avatar: dsadada!=
        },
        {
          id: 2,
          username: janusz.janek,
          firstName: Janusz,
          lastName: Janek,
          email: janusz.janek@outlook.com,
          avatar: etyerye325!=
        }
        ]
    user:
      summary: user
      value:
        {
          id: 1,
          username: jeremiasz.jaworski,
          firstName: Jeremiasz,
          lastName: Jaworski,
          role: ADMIN,
          email: jeremiasz.jaworski@outlook.com,
          country: Poland,
          city: Warszawa,
          phone: 3143141,
          zipCode: 08-500,
          avatar: 214141!=
        }
    medicalImageResponseExample:
      value:
        {
          id: 3,
          fileName: RTG_01.png,
          date: '2017-07-21T17:32:28Z',
          size: 404313,
          fileExtension: .png,
          file: 32103123=,
          diseaseProbability: 0.897979789
        }
    medicalImagesResponseExample:
      value:
        [
        {
          id: 3,
          fileName: RTG_01.png,
          date: '2017-07-21T17:32:28Z',
          size: 404313,
          fileExtension: .png,
          file: 32103123=,
          diseaseProbability: 0.43243242
        },
        {
          id: 4,
          fileName: MRI_01.jpg,
          date: '2018-07-21T17:32:28Z',
          size: 252448,
          fileExtension: .jpg,
          file: 32103123=,
          diseaseProbability: 0.96548949849
        },
        {
          id: 5,
          fileName: CT_01.dicom,
          date: '2019-07-21T17:32:28Z',
          size: 9938291,
          fileExtension: .dicom,
          file: 32103123=,
          diseaseProbability: 0.12156456456
        }
        ]
    userNotFoundExample:
      value:
        {
          message: User not found,
          status: 404,
          timestamp: 1575554106058,
          error: Not Found,
          path: /api/v1/example/path
        }

  schemas:
    User:
      type: object
      description: All user data
      required:
        - username
        - firstName
        - lastName
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - DOCTOR
            - PATIENT
        email:
          type: string
        country:
          type: string
        city:
          type: string
        phone:
          type: string
        zipCode:
          type: string
        avatar:
          type: string
    BasicUser:
      type: object
      description: Basic user data
      required:
        - username
        - firstName
        - lastName
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        avatar:
          type: string
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date
        error:
          type: string
        path:
          type: string
    MedicalImage:
      type: object
      required:
        - files
      properties:
        files:
          description: Files to upload
          type: array
          items:
            type: string
            format: binary
        description:
          description: Image description
          type: string
    MedicalImageResponse:
      type: object
      required:
        - id
        - size
        - date
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        date:
          description: Image upload date
          type: string
          format: date-time
          pattern: yyyy-MM-dd'T'HH:mm:ssZ
          example: '2017-07-21T17:32:28Z'
        size:
          description: file size in bytes
          type: integer
          format: int64
        fileExtension:
          description: file extension png,jpg,dicom
          type: string
        file:
          description: file in base64 format
          type: string
        diseaseProbability:
          description: probabilty of disease on image
          type: number
          format: double
  responses:
    UserNotFound:
      description: User Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            userNotFoundExample:
              $ref: '#/components/examples/userNotFoundExample'